# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PixelPay",
    "version": "1.0.0",
    "description": "PixelPay — AI image generation API monetized via MPP (Machine Payments Protocol). Pay-per-request with Tempo blockchain.",
    "x-guidance": "PixelPay generates images from text prompts. Send a POST to /v1/images/generate with a prompt. You will receive a 402 Payment Required with a challenge. Pay per model: Schnell $0.03, SD3.5 $0.04, Dev $0.05, Recraft $0.06, HiDream $0.08, Ideogram $0.08, Pro $0.10 USDC via Tempo blockchain, then retry with the payment credential to receive your generated image. Check GET /v1/prices for current pricing."
  },
  "x-discovery": {
    "ownershipProofs": ["dns:paid-image-api.onrender.com"]
  },
  "x-service-info": {
    "categories": ["media", "compute"],
    "docs": {
      "homepage": "https://paid-image-api.onrender.com",
      "llms": "https://paid-image-api.onrender.com/llms.txt",
      "apiReference": "https://paid-image-api.onrender.com/openapi.json"
    }
  },
  "servers": [
    {
      "url": "https://paid-image-api.onrender.com"
    }
  ],
  "paths": {
    "/v1/images/generate": {
      "post": {
        "operationId": "generateImage",
        "summary": "Generate an image from a text prompt",
        "description": "PixelPay generates images from text prompts using fal.ai models. 7 models available: Schnell $0.03, SD3.5 $0.04, Dev $0.05, Recraft $0.06, HiDream $0.08, Ideogram $0.08, Pro $0.10 USDC via MPP Tempo charge.",
        "x-payment-info": {
          "protocols": ["mpp"],
          "pricingMode": "tiered",
          "priceRange": { "min": "0.030000", "max": "0.100000" },
          "tiers": [
            { "model": "fal-ai/flux/schnell", "tier": "schnell", "price": "0.030000" },
            { "model": "fal-ai/stable-diffusion-v35-large", "tier": "sd35", "price": "0.040000" },
            { "model": "fal-ai/flux/dev", "tier": "dev", "price": "0.050000" },
            { "model": "fal-ai/recraft-v3", "tier": "recraft", "price": "0.060000" },
            { "model": "fal-ai/hidream-i1-full", "tier": "hidream", "price": "0.080000" },
            { "model": "fal-ai/ideogram/v3", "tier": "ideogram", "price": "0.080000" },
            { "model": "fal-ai/flux-pro/v1.1", "tier": "pro", "price": "0.100000" }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Text prompt describing the image to generate",
                    "example": "A serene mountain landscape at sunset with a lake reflection"
                  },
                  "model": {
                    "type": "string",
                    "description": "Model identifier. Pricing varies by model: fal-ai/flux/schnell ($0.03), fal-ai/stable-diffusion-v35-large ($0.04), fal-ai/flux/dev ($0.05), fal-ai/recraft-v3 ($0.06), fal-ai/hidream-i1-full ($0.08), fal-ai/ideogram/v3 ($0.08), fal-ai/flux-pro/v1.1 ($0.10)",
                    "default": "fal-ai/flux/schnell",
                    "enum": ["fal-ai/flux/schnell", "fal-ai/stable-diffusion-v35-large", "fal-ai/flux/dev", "fal-ai/recraft-v3", "fal-ai/hidream-i1-full", "fal-ai/ideogram/v3", "fal-ai/flux-pro/v1.1"]
                  },
                  "image_size": {
                    "type": "string",
                    "description": "Output image size",
                    "default": "landscape_4_3",
                    "enum": ["square_hd", "square", "landscape_4_3", "landscape_16_9", "portrait_4_3", "portrait_16_9"]
                  },
                  "num_images": {
                    "type": "integer",
                    "description": "Number of images to generate (1-4)",
                    "default": 1,
                    "minimum": 1,
                    "maximum": 4
                  }
                },
                "required": ["prompt"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image generated successfully",
            "headers": {
              "Payment-Receipt": {
                "description": "Base64url-encoded payment receipt",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "images": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": { "type": "string", "format": "uri" },
                          "width": { "type": "integer" },
                          "height": { "type": "integer" },
                          "content_type": { "type": "string" }
                        }
                      }
                    },
                    "prompt": { "type": "string" },
                    "model": { "type": "string" },
                    "timings": { "type": "object" }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required – respond with Authorization: Payment credential",
            "headers": {
              "WWW-Authenticate": {
                "description": "MPP payment challenge",
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    }
  }
}